Refactor line editing helpers
