+++ Makefile
@@ -6,7 +6,8 @@ MANPREFIX ?= $(PREFIX)/share/man
 SRCS := src/builtins.c src/builtins_fs.c src/builtins_jobs.c \
        src/builtins_alias.c src/builtins_func.c src/builtins_vars.c \
        src/builtins_misc.c src/execute.c src/history.c \
-       src/jobs.c src/lineedit.c src/parser.c src/lexer.c src/arith.c \
+       src/jobs.c src/lineedit.c src/history_search.c src/completion.c \
+       src/parser.c src/lexer.c src/arith.c \
        src/dirstack.c src/util.c \
        src/main.c
