+++ src/completion.h
@@ -1,0 +1,7 @@
+#ifndef COMPLETION_H
+#define COMPLETION_H
+
+void handle_completion(const char *prompt, char *buf, int *lenp, int *posp,
+                       int *disp_lenp);
+
+#endif /* COMPLETION_H */
